requirements.txt
streamlit
pyrebase4
plotly
pandas